Crate rust_i18n_support

Structs

AtomicStr

A thread-safe atomically reference-counting string.

CowStr

A wrapper for Cow<'a, str> that is specifically designed for use with the t! macro.

I18nConfig

SimpleBackend

Simple KeyValue storage backend

Constants

DEFAULT_MINIFY_KEY

The default value of minify_key feature.

DEFAULT_MINIFY_KEY_LEN

The length of auto-generated translation key

DEFAULT_MINIFY_KEY_PREFIX

The prefix of auto-generated translation key

DEFAULT_MINIFY_KEY_THRESH

The minimum length of the value to be generated the translation key

Traits

Backend

I18n backend trait

BackendExt

MinifyKey

A trait for generating translation key from a value.

Functions

is_debug

load_locales

minify_key

Generate a translation key from a value.