rust_i18n_support

Struct CowStr

pub struct CowStr<'a>(/* private fields */);

A wrapper for Cow<'a, str> that is specifically designed for use with the t! macro.

This wrapper provides additional functionality or optimizations when handling strings in the t! macro.

Implementations

impl<'a> CowStr<'a>

pub fn as_str(&self) -> &str

pub fn into_inner(self) -> Cow<'a, str>